+++ python/zef/gql/simplegql/schema_file_parser.py
@@ -61,6 +61,12 @@ def parse_partial_graphql(schema):
                 assert key in ["Algo", "JWKURL", "Audience", "Header", "Namespace",
                                "VerificationKey", "VerificationKeyEnv", "Public"], f"Unknown auth key '{key}'"
             output["auth"] = details
+        elif name == "Route":
+            assert set(keys(details)) == {"route", "hook"}
+            if "route" in ["/", "/gql"]:
+                raise Exception("Custom route should not alias '/' or '/gql'")
+            routes = output.setdefault("routes", [])
+            routes.append((details["route"], details["hook"]))
         else:
             raise Exception(f"Unsupported Zef.{name} directive")
 
@@ -244,6 +250,14 @@ def json_to_minimal_nodes(json, g):
                 vkey = os.environ[env]
             vkey = vkey.replace("\\n", '\n')
             actions += [(Z["root"], RT.AuthPresharedKey, vkey)]
+    if "routes" in json:
+        for route,hook in json["routes"]:
+            if hook not in now(g):
+                raise Exception(f"Hook named {hook} not found on schema graph")
+            actions += [(Z["root"], RT.Route,
+                         {ET.Route: {RT.Route: route,
+                                     RT.Hook: g | now | get[hook] | collect}})]
+                
 
     for gql_name,typ in core_types.items():
         actions += [